+++ src/main/java/com/nedacort/challengespringbackend/persistense/repositoryjpa/GenderJpaRepository.java
@@ -3,5 +3,5 @@ package com.nedacort.challengespringbackend.persistense.repositoryjpa;
 import com.nedacort.challengespringbackend.persistense.entities.Gender;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GenderJpaRepository extends CrudRepository<Gender, Integer> {
+public interface GenderJpaRepository extends CrudRepository<Gender, Long> {
 }
