+++ src/main/java/com/nedacort/challengespringbackend/persistense/repositoryjpa/MovieJpaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MovieJpaRepository extends JpaRepository<Movie, Integer> {
+public interface MovieJpaRepository extends JpaRepository<Movie, Long> {
 
 //
 //    @Query(value = "select peliculas.id,peliculas.imagen, peliculas.titulo, peliculas.fecha_creacion,\n" +
